demobatdongsan v1

+++ App/views/Admin/ListNewsView.php
@@ -95,33 +95,43 @@
             <img src="<?php echo WEB_ROOT;?>/public/admin/img/basic/logo-white.png" alt="">
         </div>
         <ul class="sidebar-menu">
-            <li class="header"><strong>MAIN NAVIGATION</strong></li>
+            <li class="header"><strong>QUẢN LÝ</strong></li>
             <li class="treeview">
                 <a href="<?php echo WEB_ROOT;?>">
-                    <i class="icon icon-sailing-boat-water purple-text s-18"></i> <span>View Page</span>
+                    <i class="icon icon-sailing-boat-water purple-text s-18"></i> <span>Xem trang web</span>
                 </a>
             </li>
             <li class="treeview"><a href="#">
-                <i class="icon icon icon-package blue-text s-18"></i>
-                <span>Products</span>
+                <i class="icon icon icon-package blue-text s-18"></i>Nhà <i class="icon icon-angle-left s-18 pull-right"></i>
             </a>
                 <ul class="treeview-menu">
-                    <li><a href="<?php echo WEB_ROOT;?>/admin/"><i class="icon icon-circle-o"></i>All Products</a>
+                    <li><a href="<?php echo WEB_ROOT;?>/admin/"><i class="icon icon-circle-o"></i>Tất cả nhà</a>
                     </li>
-                    <li><a href="<?php echo WEB_ROOT;?>/admin/addProduct/"><i class="icon icon-add"></i>Add New </a>
+                    <li><a href="<?php echo WEB_ROOT;?>/admin/addProduct/"><i class="icon icon-add"></i>Thêm nhà mới</a>
                     </li>
                 </ul>
             </li>
-
+            <li class="treeview">
+                <a href="#">
+                    <i class="icon icon icon-house blue-text s-18"></i>
+                    Chung cư <i class="icon icon-angle-left s-18 pull-right"></i>
+                </a>
+                <ul class="treeview-menu">
+                    <li><a href="<?php echo WEB_ROOT;?>/admin/getListApart/"><i class="icon icon-circle-o"></i>Chung cư Landmark</a>
+                    </li>
+                    <li><a href="<?php echo WEB_ROOT;?>/admin/addApart/"><i class="icon icon-add"></i>Thêm Chung cư </a>
+                    </li>
+                </ul>
+            </li>
             <?php
                 if(isset($_SESSION['role']) && $_SESSION['role'] == '1'){
                     echo '
-                        <li class="treeview"><a href="#"><i class="icon icon-account_box light-green-text s-18"></i>Users<i
+                        <li class="treeview"><a href="#"><i class="icon icon-account_box light-green-text s-18"></i>Người dùng<i
                         class="icon icon-angle-left s-18 pull-right"></i></a>
                             <ul class="treeview-menu">
-                                <li><a href="'.WEB_ROOT.'/admin/listUser/"><i class="icon icon-circle-o"></i>All Users</a>
+                                <li><a href="'.WEB_ROOT.'/admin/listUser/"><i class="icon icon-circle-o"></i>Tất cả người dùng</a>
                                 </li>
-                                <li><a href="'.WEB_ROOT.'/admin/addUser/"><i class="icon icon-add"></i>Add User</a>
+                                <li><a href="'.WEB_ROOT.'/admin/addUser/"><i class="icon icon-add"></i>Thêm người dùng mới</a>
                                 </li>
                             </ul>
                         </li>
@@ -131,12 +141,12 @@
                             <span>Chờ duyệt</span>
                         </a>
 
-                        <li class="treeview"><a href="#"><i class="icon icon-documents3 light-green-text s-18"></i>News<i
+                        <li class="treeview"><a href="#"><i class="icon icon-documents3 light-green-text s-18"></i>Tin tức<i
                         class="icon icon-angle-left s-18 pull-right"></i></a>
                             <ul class="treeview-menu">
-                                <li><a href="'.WEB_ROOT.'/admin/News/"><i class="icon icon-circle-o"></i>All News</a>
+                                <li><a href="'.WEB_ROOT.'/admin/listNews/"><i class="icon icon-circle-o"></i>Tất cả tin tức</a>
                                 </li>
-                                <li><a href="'.WEB_ROOT.'/admin/addNews/"><i class="icon icon-add"></i>Add News</a>
+                                <li><a href="'.WEB_ROOT.'/admin/addNews/"><i class="icon icon-add"></i>Thêm tin tức mới</a>
                                 </li>
                             </ul>
                         </li>
@@ -347,7 +357,7 @@
                                             <td>
                                                 <a href="'.WEB_ROOT.'/admin/detailNews/?n_id='.$value['n_id'].'"><span class="r-3 badge badge-success">Xem</span></a>
                                                 <a href="'.WEB_ROOT.'/admin/editNews/?n_id='.$value['n_id'].'"><span class="r-3 badge badge-warning">Sửa</span></a>
-                                                <a href="'.WEB_ROOT.'/admin/deleteNews/?n_id='.$value['n_id'].'"><span class="r-3 badge badge-danger">xóa</span></a>
+                                                <a href="'.WEB_ROOT.'/admin/deleteNews/?n_id='.$value['n_id'].'" onclick="return confirm(\'BẠN CÓ CHẮC MUỐN XÓA ?\')"><span class="r-3 badge badge-danger">xóa</span></a>
                                             </td>
                                         </tr>
                                         ';
